Refactor, tidy and comment symbolic model class storage.
symbolic.model.Model remains the top-level interface,
but contains only the read-only access methods for models.
Some method implementations are moved there as defaults.

New class ModelSymbolic is the (mutable) base implementation,
covering features common to all models, e.g., initial states,
reachable states, labels. Most of this was in ProbModel
previously. The split between Model/ModelSymbolic mirrors the
situation in the explicit engine with Model and ModelExplicit.

ProbModel (DTMC) and NondetModel (MDP) now separately extend
ModelSymbolic and StochModel (CTMC) extends ProbModel.

Lots of tidying and commenting.

A few unimplemented/deprecated methods are removed.
